feat: Make store state simpler
Ideally, having an inner `TVar` in our in-memory key-val store ought to
increase our ability for concurrent access, and indeed it does, but not
for a common use case.

You see, the approach of having a `HashMap k (TVar v)` makes it such
that modifications to `v` can be made concurrently without conflicts
when other threads are modifiying other `v`s. However this doesn't stop
all threads from retrying should any changes be made to the underlying
HashMap itself which tends to happen when a new element is added or
removed from the store. Reads aren't affected either way

So, having a `v` wrapped in a `TVar` only offers us more concurrent
acess granularity when our workload is fairly "update" heavy which a
program like Redis is not. Key value stores like redis AFAIU are very
"set and delete" heavy thus rendering moot the supposed concurrency gains of
an inner `TVar v` in our in-memory key-val store representation (the
HashMap)

In response to this realization, I opted to make things simpler by only
wrapping the HashMap itself (the store) in a `TVar`. We _could_ use an
`IORef` instead now, but that would be a more pervasive change though
with potential performance gains (due to a lower overhead relative to
STM). However, I still want to play with STM so I won't be moving over to an `IORef` just yet